requirements.txt
# FastAPI specific requirements
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.1.1

# Core ML/AI dependencies
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.0
langchain-groq==0.3.5
langchain-huggingface==0.3.0

# Vector database and embeddings
faiss-cpu==1.11.0
sentence-transformers==5.0.0

# Document processing
pypdf==5.7.0
pymupdf
python-docx==1.1.2

# API clients
groq==0.29.0
openai>=1.0.0

# Additional utilities
numpy>=1.24.0
requests>=2.31.0



# RAG Evaluation Dependencies

scikit-learn==1.7.0
nltk==3.9.1

streamlit
plotly

# Environment and utility packages

glob2  # enhanced glob functionality (optional, glob is built-in)

# Text processing
tiktoken  # for token counting (used by OpenAI embeddings)


pandas

# For development and debugging (optional)
jupyter
ipython




google-cloud-documentai
google-api-core
PyPDF2
openpyxl
reportlab
docx2pdf
python-pptx
Pillow
aiofiles
motor
langdetect